+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [101.12.2](https://github.com/dhis2/maps-app/compare/v101.12.1...v101.12.2) (2026-06-11)
+
+
+### Bug Fixes
+
+* prevent timeline crash on load failure and fix period after drilling up/down [DHIS2-19063] [DHIS2-21113] ([#3664](https://github.com/dhis2/maps-app/issues/3664)) ([e9bd0c3](https://github.com/dhis2/maps-app/commit/e9bd0c34e2c1c4952ec35497bc8eb94b7e9f59f7))
+
+## [101.12.1](https://github.com/dhis2/maps-app/compare/v101.12.0...v101.12.1) (2026-06-11)
+
+
+### Bug Fixes
+
+* preserve saved dates and default period when editing/adding thematic and event layers [DHIS2-21516] [DHIS2-21517] ([#3666](https://github.com/dhis2/maps-app/issues/3666)) ([2dd021b](https://github.com/dhis2/maps-app/commit/2dd021b91c3b5b2c0d14c6da108cdf39ffbdecbf))
+
+# [101.12.0](https://github.com/dhis2/maps-app/compare/v101.11.1...v101.12.0) (2026-06-11)
+
+
+### Features
+
+* enhance map legends, classification, and styling [DHIS2-18242] ([#3661](https://github.com/dhis2/maps-app/issues/3661)) ([c2b725d](https://github.com/dhis2/maps-app/commit/c2b725d417c20b17f3a3be98915834862aecd746))
+
 ## [101.11.1](https://github.com/dhis2/maps-app/compare/v101.11.0...v101.11.1) (2026-06-01)